apollo queries, mutations, and resolvers interacting with flutter app now

+++ apollo/src/Coupon.ts
@@ -8,7 +8,7 @@ const couponSchema = new mongoose.Schema({
     event: [{ type: Schema.Types.ObjectId, ref: 'Event' }],    
     users: [{ type: Schema.Types.ObjectId, ref: 'User' }],        
     accessCoupon: Boolean,
-    capacity: Int32Array,
+    capacity: BigInt,
 });
 
 const Coupon = mongoose.model('Coupon', couponSchema);
